ai-generator/requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone-client
langchain-pinecone
langchain
langchain-huggingface
google-generativeai
baml-py
python-dotenv